pyproject.toml
[project]
name = "rbchat-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "langchain",
    "langgraph",
    "python-dotenv",
    "psycopg2-binary",
    "langchain-mcp-adapters",
    "fastmcp",
    "langchain-openai",
]
